% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bfsnr.R
\name{add_bfsnr}
\alias{add_bfsnr}
\title{add bfsnr to geocoordinates}
\usage{
add_bfsnr(df, crs, datum = "latest")
}
\arguments{
\item{df}{dataframe containing the coordinates which should be mapped to a bfsnr}

\item{crs}{epsg number of the projection of the dataset}

\item{datum}{datum as string, used to define which data should be used for the matching}
}
\description{
add bfsnr to geocoordinates
}
