% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_shapes.R
\name{download_gem_grenzen}
\alias{download_gem_grenzen}
\title{Function to download the swissboundary shapes}
\usage{
download_gem_grenzen(dest_dir, which_date = "latest")
}
\arguments{
\item{dest_dir}{path to destination}

\item{which_date}{which date should be downloadedz, default is: "latest"}
}
\description{
Function to download the swissboundary shapes
}
